+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Version 0.2.0
+
+- Defining Memory Pointer for managing memory allocation in WASM's linear memory.
+- Implementing `deallocate` API to free the allocated memory
+
 ## Version 0.1.0
 
 - Instantiating contract actor